+++ core/transplantlistmodel.cpp
@@ -14,6 +14,8 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+#include <QDate>
+
 #include "transplantlistmodel.h"
 
 TransplantListModel::TransplantListModel(QObject *parent, const QString &tableName)
@@ -26,4 +28,35 @@ bool TransplantListModel::filterAcceptsRow(int sourceRow, const QModelIndex &sou
 {
     int year = rowValue(sourceRow, sourceParent, "year").toInt();
     return (year == m_year) && SortFilterProxyModel::filterAcceptsRow(sourceRow, sourceParent);
+}
+
+bool TransplantListModel::lessThan(const QModelIndex &left, const QModelIndex &right) const
+{
+    if (m_sortColumn == QStringLiteral("planting_date")) {
+        auto lhs = fieldDate(left.row(), left.parent(), "planting_date");
+        auto rhs = fieldDate(right.row(), right.parent(), "planting_date");
+        return lhs < rhs;
+    } else if (m_sortColumn == QStringLiteral("crop")) {
+        auto leftCrop = rowValue(left.row(), left.parent(), "crop").toString();
+        auto rightCrop = rowValue(right.row(), right.parent(), "crop").toString();
+        int cmp = leftCrop.localeAwareCompare(rightCrop);
+        if (cmp < 0) {
+            return true;
+        } else if (cmp == 0) {
+            auto leftVariety = rowValue(left.row(), left.parent(), "variety").toString();
+            auto rightVariety = rowValue(right.row(), right.parent(), "variety").toString();
+            return leftVariety.localeAwareCompare(rightVariety) < 0;
+        }
+    } else if (m_sortColumn == QStringLiteral("variety")
+               || m_sortColumn == QStringLiteral("seed_company")) {
+        auto leftVariety = rowValue(left.row(), left.parent(), m_sortColumn).toString();
+        auto rightVariety = rowValue(right.row(), right.parent(), m_sortColumn).toString();
+        return leftVariety.localeAwareCompare(rightVariety) < 0;
+    } else if (m_sortColumn == QStringLiteral("seeds_number")
+               || m_sortColumn == QStringLiteral("seeds_quantity")) {
+        auto lhs = rowValue(left.row(), left.parent(), m_sortColumn).toInt();
+        auto rhs = rowValue(right.row(), right.parent(), m_sortColumn).toInt();
+        return lhs < rhs;
+    }
+    return SortFilterProxyModel::lessThan(left, right);
 }