+++ core/transplantlistmodel.h
@@ -31,6 +31,7 @@ public:
 
 protected:
     bool filterAcceptsRow(int sourceRow, const QModelIndex &sourceParent) const override;
+    bool lessThan(const QModelIndex &left, const QModelIndex &right) const override;
 };
 
 #endif // TRANSPLANTLISTMODEL_H